+++ flow-typed/moduleHotReload.js
@@ -1,7 +1,0 @@
-/* eslint-disable no-unused-vars */
-
-declare var module: {
-  hot: {
-    accept(path: string, callback: () => void): void,
-  },
-};