+++ src/front/hoc/withEnterAnimation/__tests__/withEnterAnimation.test.js
@@ -1,18 +1,15 @@
 // @flow
 
-// #region imports
 import React from 'react';
-import renderer from 'react-test-renderer'; // needed both for snpashot testing but also to prevent errors from enzyme
+import { shallow } from 'enzyme';
 import withEnterAnimation from '../withEnterAnimation';
-// #endregion
 
 describe('withEnterAnimation HOC', () => {
   it('renders as expected', () => {
     const DummyComponent = () => <p>component</p>;
+    const WrappedComponent = withEnterAnimation()(DummyComponent);
 
-    const component = renderer
-      .create(withEnterAnimation()(DummyComponent))
-      .toJSON();
+    const component = shallow(<WrappedComponent />);
     expect(component).toMatchSnapshot();
   });
 });