+++ src/front/redux/modules/reducers.js
@@ -1,10 +1,8 @@
 // @flow
 
-// #region imports
 import { combineReducers } from 'redux';
 import fakeModuleWithFetch from './fakeModuleWithFetch';
 import userAuth from './userAuth';
-// #endregion
 
 export const reducers = {
   fakeModuleWithFetch,