+++ src/front/routes/routes.js
@@ -1,8 +1,6 @@
 // @flow
 
-// #region imports
 import loadable from 'loadable-components';
-// #endregion
 
 export const Home = loadable(() => import('../pages/home'));
 export const About = loadable(() => import('../pages/about'));