+++ src/front/style/GlobalStyles.js
@@ -1,8 +1,6 @@
 // @flow
 
-// #region imports
-import { createGlobalStyle } from 'styled-components'
-// #endregion
+import { createGlobalStyle } from 'styled-components';
 
 const GlobalStyle = createGlobalStyle`
   html, body {