+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# v7.0.0
+- migrate front to Typescript
+
 # v6.1.1
 - upgrade `styled-components` to v4+ (*now uses createGlobalStyle()*)
 - upgrade `loadable-components` to 2.2.3 (*max version compatible with reactsnap, avoiding flashing on application start*)